requirements.txt
# Web Framework
streamlit>=1.28.0

# NLP & ML
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Vector Database
faiss-cpu>=1.7.0

# Document Processing
pypdf>=3.0.0
python-docx>=1.0.0

# Search & Ranking
rank-bm25>=0.2.2

# Data Processing
numpy>=1.24.0
scikit-learn>=1.3.0

# Optional: for better text processing
# nltk>=3.8.0